Speed optimization, fixed bug in food eating.

+++ src/pl/edu/uj/tcs/kuini/model/live/ILivePlayer.java
@@ -3,7 +3,7 @@ package pl.edu.uj.tcs.kuini.model.live;
 import pl.edu.uj.tcs.kuini.model.IPlayer;
 
 public interface ILivePlayer extends IPlayer {
-	void changeScore(long change);
+	void changeScore(float change);
 	boolean shouldAttack(ILivePlayer player);
 	void changeFood(float change);
 }
